+++ src/main/java/com/daml/extensions/damlmavenplugin/MojoBase.java
@@ -4,12 +4,19 @@
  */
 package com.daml.extensions.damlmavenplugin;
 
+import java.io.File;
 import java.io.InputStream;
+import java.nio.file.Paths;
 import java.util.Scanner;
 
 import org.apache.maven.plugin.AbstractMojo;
+import org.apache.maven.plugin.MojoFailureException;
+import org.apache.maven.plugins.annotations.Parameter;
 
 public abstract class MojoBase extends AbstractMojo {
+
+    @Parameter(defaultValue = "${project.basedir}")
+    protected File damlProjectDirectory;
 
     protected void redirectOutput(InputStream in) {
         try (Scanner scanner = new Scanner(in)) {
@@ -17,5 +24,16 @@ public abstract class MojoBase extends AbstractMojo {
                 getLog().info(scanner.nextLine());
             }
         }
+    }
+
+    public File getDamlProjectDirectory() {
+        if (damlProjectDirectory == null) {
+            return Paths.get("").toFile();
+        } else
+            return damlProjectDirectory;
+    }
+
+    protected DamlProject createDamlProject() throws MojoFailureException {
+        return DamlProject.create();
     }
 }